add item function to firebase
